feat: Map location implemented on detail and implement SingleLiveEvent to avoid unnecesary emits

+++ data/src/main/java/com/idealista/data/dto/ad_detail/AdDetailResponse.kt
@@ -36,8 +36,8 @@ data class ImageDetailResponse(
 )
 
 data class UbicationDetailResponse(
-    val latitude: Double,
-    val longitude: Double
+    val latitude: Float,
+    val longitude: Float
 )
 
 data class MoreCharacteristicsDetailResponse(
